Merge instance and device extensions for drm external images
Even though the Vulkan spec recommends splitting this extension, a split
results in two very awkward, seemingly incomplete extensions.  The
features exposed here are most naturally expressed as a single
extension. Therefore, for convenience, now let's keep this as a single
spec for now and revisit this issue before finalizing the extension
spec.