Split and rename VK_CHROMIUM_external_drm_image
The extension defined how to import images but not export them. When the
export feature arrives, the import and export should be defined by
separate extension. So now is a good time to split the base
functionality and the import functionality of
VK_CHROMIUM_external_drm_image into separate extensions.

This extension is closely related to VK_CHROMIUM_dma_buf_memory_import
and VK_CHROMIUM_dma_buf_memory_export, so rename to match those
extensions.

The result:
    VK_CHROMIUM_external_drm_image
        -> VK_CHROMIUM_dma_buf_image_format_properties
        -> VK_CHROMIUM_dma_buf_image_import